Combine LayoutTable into LayoutPanel component
